+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 	resources.tertiary.value
 - applyChatCardDamage should be working. Right Click on Roll Chat messages.
 - fixed issue with scrollbar that was only vissable within browser, and not Electron client.
+- added custom encumbrance options
 
 ## Version 0.1.3
 - Added an option for Game Masters to preview and show items to parties groups